move font renderer
