+++ src/main/java/com/skidders/sigma/utils/render/font/data/CharacterData.java
@@ -1,4 +1,4 @@
-package com.skidders.sigma.utils.font.data;
+package com.skidders.sigma.utils.render.font.data;
 
 import org.lwjgl.opengl.GL11;
